# api/http/oas_router_gen.go
// Code generated by ogen, DO NOT EDIT.

package api

import (
	"net/http"
	"net/url"
	"strings"

	"github.com/ogen-go/ogen/uri"
)

var (
	rn3AllowedHeaders = map[string]string{
		"POST": "Authorization",
	}
	rn5AllowedHeaders = map[string]string{
		"GET": "Authorization",
	}
	rn7AllowedHeaders = map[string]string{
		"POST": "Authorization",
	}
	rn9AllowedHeaders = map[string]string{
		"POST": "Content-Type",
	}
	rn10AllowedHeaders = map[string]string{
		"POST": "Content-Type",
	}
	rn12AllowedHeaders = map[string]string{
		"POST": "Content-Type",
	}
	rn14AllowedHeaders = map[string]string{
		"DELETE": "Authorization",
	}
	rn15AllowedHeaders = map[string]string{
		"POST": "Content-Type",
	}
)

func (s *Server) cutPrefix(path string) (string, bool) {
	prefix := s.cfg.Prefix
	if prefix == "" {
		return path, true
	}
	if !strings.HasPrefix(path, prefix) {
		// Prefix doesn't match.
		return "", false
	}
	// Cut prefix from the path.
	return strings.TrimPrefix(path, prefix), true
}

// ServeHTTP serves http request as defined by OpenAPI v3 specification,
// calling handler that matches the path or returning not found error.
func (s *Server) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	elem := r.URL.Path
	elemIsEscaped := false
	if rawPath := r.URL.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
			elemIsEscaped = strings.ContainsRune(elem, '%')
		}
	}

	elem, ok := s.cutPrefix(elem)
	if !ok || len(elem) == 0 {
		s.notFound(w, r)
		return
	}
	args := [1]string{}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/"

			if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'a': // Prefix: "a"

				if l := len("a"); len(elem) >= l && elem[0:l] == "a" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'd': // Prefix: "dmin/users/"

					if l := len("dmin/users/"); len(elem) >= l && elem[0:l] == "dmin/users/" {
						elem = elem[l:]
					} else {
						break
					}

					// Param: "userId"
					// Match until "/"
					idx := strings.IndexByte(elem, '/')
					if idx < 0 {
						idx = len(elem)
					}
					args[0] = elem[:idx]
					elem = elem[idx:]

					if len(elem) == 0 {
						break
					}
					switch elem[0] {
					case '/': // Prefix: "/revoke-chain"

						if l := len("/revoke-chain"); len(elem) >= l && elem[0:l] == "/revoke-chain" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch r.Method {
							case "POST":
								s.handleAdminUsersUserIdRevokeChainPostRequest([1]string{
									args[0],
								}, elemIsEscaped, w, r)
							default:
								s.notAllowed(w, r, notAllowedParams{
									allowedMethods: "POST",
									allowedHeaders: rn3AllowedHeaders,
									acceptPost:     "",
									acceptPatch:    "",
								})
							}

							return
						}

					}

				case 'u': // Prefix: "uth/verify"

					if l := len("uth/verify"); len(elem) >= l && elem[0:l] == "uth/verify" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "GET":
							s.handleAuthVerifyGetRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "GET",
								allowedHeaders: rn5AllowedHeaders,
								acceptPost:     "",
								acceptPatch:    "",
							})
						}

						return
					}

				}

			case 'i': // Prefix: "invites"

				if l := len("invites"); len(elem) >= l && elem[0:l] == "invites" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "POST":
						s.handleInvitesPostRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "POST",
							allowedHeaders: rn7AllowedHeaders,
							acceptPost:     "",
							acceptPatch:    "",
						})
					}

					return
				}

			case 'l': // Prefix: "login"

				if l := len("login"); len(elem) >= l && elem[0:l] == "login" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch r.Method {
					case "POST":
						s.handleLoginPostRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "POST",
							allowedHeaders: rn9AllowedHeaders,
							acceptPost:     "application/json",
							acceptPatch:    "",
						})
					}

					return
				}
				switch elem[0] {
				case '/': // Prefix: "/begin"

					if l := len("/begin"); len(elem) >= l && elem[0:l] == "/begin" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "POST":
							s.handleLoginBeginPostRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "POST",
								allowedHeaders: nil,
								acceptPost:     "",
								acceptPatch:    "",
							})
						}

						return
					}

				}

			case 'r': // Prefix: "registration/"

				if l := len("registration/"); len(elem) >= l && elem[0:l] == "registration/" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'b': // Prefix: "begin"

					if l := len("begin"); len(elem) >= l && elem[0:l] == "begin" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "POST":
							s.handleRegistrationBeginPostRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "POST",
								allowedHeaders: rn10AllowedHeaders,
								acceptPost:     "application/json",
								acceptPatch:    "",
							})
						}

						return
					}

				case 'c': // Prefix: "complete"

					if l := len("complete"); len(elem) >= l && elem[0:l] == "complete" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "POST":
							s.handleRegistrationCompletePostRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "POST",
								allowedHeaders: rn12AllowedHeaders,
								acceptPost:     "application/json",
								acceptPatch:    "",
							})
						}

						return
					}

				}

			case 's': // Prefix: "sessions/"

				if l := len("sessions/"); len(elem) >= l && elem[0:l] == "sessions/" {
					elem = elem[l:]
				} else {
					break
				}

				// Param: "sessionId"
				// Leaf parameter, slashes are prohibited
				idx := strings.IndexByte(elem, '/')
				if idx >= 0 {
					break
				}
				args[0] = elem
				elem = ""

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "DELETE":
						s.handleSessionsSessionIdDeleteRequest([1]string{
							args[0],
						}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "DELETE",
							allowedHeaders: rn14AllowedHeaders,
							acceptPost:     "",
							acceptPatch:    "",
						})
					}

					return
				}

			case 't': // Prefix: "token/refresh"

				if l := len("token/refresh"); len(elem) >= l && elem[0:l] == "token/refresh" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "POST":
						s.handleTokenRefreshPostRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "POST",
							allowedHeaders: rn15AllowedHeaders,
							acceptPost:     "application/json",
							acceptPatch:    "",
						})
					}

					return
				}

			}

		}
	}
	s.notFound(w, r)
}

// Route is route object.
type Route struct {
	name           string
	summary        string
	operationID    string
	operationGroup string
	pathPattern    string
	count          int
	args           [1]string
}

// Name returns ogen operation name.
//
// It is guaranteed to be unique and not empty.
func (r Route) Name() string {
	return r.name
}

// Summary returns OpenAPI summary.
func (r Route) Summary() string {
	return r.summary
}

// OperationID returns OpenAPI operationId.
func (r Route) OperationID() string {
	return r.operationID
}

// OperationGroup returns the x-ogen-operation-group value.
func (r Route) OperationGroup() string {
	return r.operationGroup
}

// PathPattern returns OpenAPI path.
func (r Route) PathPattern() string {
	return r.pathPattern
}

// Args returns parsed arguments.
func (r Route) Args() []string {
	return r.args[:r.count]
}

// FindRoute finds Route for given method and path.
//
// Note: this method does not unescape path or handle reserved characters in path properly. Use FindPath instead.
func (s *Server) FindRoute(method, path string) (Route, bool) {
	return s.FindPath(method, &url.URL{Path: path})
}

// FindPath finds Route for given method and URL.
func (s *Server) FindPath(method string, u *url.URL) (r Route, _ bool) {
	var (
		elem = u.Path
		args = r.args
	)
	if rawPath := u.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
		}
		defer func() {
			for i, arg := range r.args[:r.count] {
				if unescaped, err := url.PathUnescape(arg); err == nil {
					r.args[i] = unescaped
				}
			}
		}()
	}

	elem, ok := s.cutPrefix(elem)
	if !ok {
		return r, false
	}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/"

			if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'a': // Prefix: "a"

				if l := len("a"); len(elem) >= l && elem[0:l] == "a" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'd': // Prefix: "dmin/users/"

					if l := len("dmin/users/"); len(elem) >= l && elem[0:l] == "dmin/users/" {
						elem = elem[l:]
					} else {
						break
					}

					// Param: "userId"
					// Match until "/"
					idx := strings.IndexByte(elem, '/')
					if idx < 0 {
						idx = len(elem)
					}
					args[0] = elem[:idx]
					elem = elem[idx:]

					if len(elem) == 0 {
						break
					}
					switch elem[0] {
					case '/': // Prefix: "/revoke-chain"

						if l := len("/revoke-chain"); len(elem) >= l && elem[0:l] == "/revoke-chain" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch method {
							case "POST":
								r.name = AdminUsersUserIdRevokeChainPostOperation
								r.summary = "Каскадный отзыв сессий по цепочке приглашений"
								r.operationID = ""
								r.operationGroup = ""
								r.pathPattern = "/admin/users/{userId}/revoke-chain"
								r.args = args
								r.count = 1
								return r, true
							default:
								return
							}
						}

					}

				case 'u': // Prefix: "uth/verify"

					if l := len("uth/verify"); len(elem) >= l && elem[0:l] == "uth/verify" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "GET":
							r.name = AuthVerifyGetOperation
							r.summary = "Проверить access token"
							r.operationID = "AuthVerifyGet"
							r.operationGroup = ""
							r.pathPattern = "/auth/verify"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

				}

			case 'i': // Prefix: "invites"

				if l := len("invites"); len(elem) >= l && elem[0:l] == "invites" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch method {
					case "POST":
						r.name = InvitesPostOperation
						r.summary = "Создать инвайт-код"
						r.operationID = ""
						r.operationGroup = ""
						r.pathPattern = "/invites"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}

			case 'l': // Prefix: "login"

				if l := len("login"); len(elem) >= l && elem[0:l] == "login" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch method {
					case "POST":
						r.name = LoginPostOperation
						r.summary = "Завершить вход"
						r.operationID = ""
						r.operationGroup = ""
						r.pathPattern = "/login"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}
				switch elem[0] {
				case '/': // Prefix: "/begin"

					if l := len("/begin"); len(elem) >= l && elem[0:l] == "/begin" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "POST":
							r.name = LoginBeginPostOperation
							r.summary = "Начать вход"
							r.operationID = ""
							r.operationGroup = ""
							r.pathPattern = "/login/begin"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

				}

			case 'r': // Prefix: "registration/"

				if l := len("registration/"); len(elem) >= l && elem[0:l] == "registration/" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'b': // Prefix: "begin"

					if l := len("begin"); len(elem) >= l && elem[0:l] == "begin" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "POST":
							r.name = RegistrationBeginPostOperation
							r.summary = "Начать регистрацию по инвайт-коду"
							r.operationID = ""
							r.operationGroup = ""
							r.pathPattern = "/registration/begin"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

				case 'c': // Prefix: "complete"

					if l := len("complete"); len(elem) >= l && elem[0:l] == "complete" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "POST":
							r.name = RegistrationCompletePostOperation
							r.summary = "Завершить регистрацию"
							r.operationID = ""
							r.operationGroup = ""
							r.pathPattern = "/registration/complete"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

				}

			case 's': // Prefix: "sessions/"

				if l := len("sessions/"); len(elem) >= l && elem[0:l] == "sessions/" {
					elem = elem[l:]
				} else {
					break
				}

				// Param: "sessionId"
				// Leaf parameter, slashes are prohibited
				idx := strings.IndexByte(elem, '/')
				if idx >= 0 {
					break
				}
				args[0] = elem
				elem = ""

				if len(elem) == 0 {
					// Leaf node.
					switch method {
					case "DELETE":
						r.name = SessionsSessionIdDeleteOperation
						r.summary = "Выйти с конкретного устройства"
						r.operationID = ""
						r.operationGroup = ""
						r.pathPattern = "/sessions/{sessionId}"
						r.args = args
						r.count = 1
						return r, true
					default:
						return
					}
				}

			case 't': // Prefix: "token/refresh"

				if l := len("token/refresh"); len(elem) >= l && elem[0:l] == "token/refresh" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch method {
					case "POST":
						r.name = TokenRefreshPostOperation
						r.summary = "Обновить access-токен"
						r.operationID = ""
						r.operationGroup = ""
						r.pathPattern = "/token/refresh"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}

			}

		}
	}
	return r, false
}